backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiosqlite==0.20.0
greenlet==3.0.3
supabase==1.2.0
gotrue==1.3.1
postgrest==0.10.8
realtime==1.0.0
storage3==0.6.0
supafunc==0.2.3
httpx==0.24.1
python-dotenv==1.0.1
aiofiles
